requirements.txt
Flask
pandas
numpy
openai
scikit-learn
PyYaml